Move helpers to SetupHelpers.hs
